fix: gas pricing and re-use of deployed facets
Fixed a few issues with parsing and configuring gas price and gas limit
options.

Added the ability to save the facets out as they are deployed, so that
they can be re-used in subsequent new Diamond deployments. This is
especially useful when there are issues calling diamondCut *after*
having paid for facet deployment.